+++ conf/user.json
@@ -1,0 +1,149 @@
+//  使用 json5 标准格式，https://json5.org/
+[
+  {
+    "welcome": "用户一，这是一条欢迎语，每次打卡前会打印这句话，用来标记当前打卡用户，如：正在为 *** 打卡......",
+    "phone": "**",  // 完美校园登录账号
+    "password": "**",  // 完美校园登录密码
+    "device_id": "**",  // 已验证完美校园登录设备ID
+    "healthy_checkin": {
+      "one_check": {  // 第一类健康打卡
+        "enable": false,  // true 为打开，false 为关闭
+        // 以下为数据修改，如果数据获取出现问题，请抓包修改对应项
+        "post_json": {
+          "areaStr": "",
+          "deptStr": "",
+          "deptid": "",
+          "stuNo": "",
+          "username": "",
+          "userid": "",
+          "updatainfo": []
+        }
+      },
+      "two_check": {  // 第二类健康打卡
+        "enable": false,  // true 为打开，false 为关闭
+        // 以下为数据修改，如果数据获取出现问题，请抓包修改或添加
+        "post_json":{
+          "address": "",
+          "stuNo": "",
+          //  百度坐标拾取：http://api.map.baidu.com/lbsapi/getpoint/index.html，可用来获取下面两个值
+          "longitude": "",  // 请对应 address 填写该值，否则老师那边会显示手动签到
+          "latitude": ""  // 请对应 address 填写该值，否则老师那边会显示手动签到
+        }
+      }
+    },
+    "campus_checkin": {  // 校内打卡
+      "enable": false,  // true 为打开，false 为关闭
+      "post_json": {  // 以下为数据修改，如果数据获取出现问题，请抓包修改对应项
+          "areaStr": "",
+          "deptStr": "",
+          "deptid": "",
+          "stuNo": "",
+          "username": "",
+          "userid": "",
+          "updatainfo": [
+            {
+              "propertyname": "temperature",
+              "value": "36.4"
+            },
+            {
+              "propertyname": "symptom",
+              "value": "无症状"
+            },
+          ]
+        }
+    },
+    "push": {  // 单人推送设置，若全部关闭，则使用 push.json 文件的配置，进行统一推送
+      "email": {  //  自定义邮箱推送，使用 qq 邮箱推送，就用 qq 邮箱的 smtp 服务地址和端口
+        "enable": false,  // true 为打开，false 为关闭
+        "smtp_address": "smtp.qq.com",  // stmp服务地址
+        "smtp_port": 465,  // stmp服务端口
+        "send_email": "***@qq.com",  // 发送邮箱的邮箱地址
+        "send_pwd": "****",  // 发送邮箱的邮箱授权码
+        "receive_email": "**@qq.com"  // 接收信息的邮箱地址
+      },
+      "wechat": {  //  Server酱·Turbo版：https://sct.ftqq.com/
+        "enable": false,  // true 为打开，false 为关闭
+        "send_key": "**"  // Server 酱推送密钥
+      },
+      "qmsg": {  //  Qmsg酱：https://qmsg.zendee.cn/index.html
+        "enable": false,  // true 为打开，false 为关闭
+        "key": "**",  // 推送所必须的 key
+        "type": "send",  //  send 私聊消息推送，group 群消息推送
+        "qq_num": "**"  // 上面设置send，则填写接受推送的qq号，group则填写群号，根据官网要求群号需要邮箱申请才能推送
+      }
+    }
+  },
+  {
+    "welcome": "用户二，这是一条欢迎语，每次打卡前会打印这句话，用来标记当前打卡用户，如：正在为 *** 打卡......",
+    "phone": "**",  // 完美校园登录账号
+    "password": "**",  // 完美校园登录密码
+    "device_id": "***",  // 已验证完美校园登录设备ID
+    "healthy_checkin": {
+      "one_check": {  // 第一类健康打卡
+        "enable": false,  // true 为打开，false 为关闭
+        // 以下为数据修改，如果数据获取出现问题，请抓包修改对应项
+        "post_json": {
+          "areaStr": "",
+          "deptStr": "",
+          "deptid": "",
+          "stuNo": "",
+          "username": "",
+          "userid": "",
+          "updatainfo": []
+        }
+      },
+      "two_check": {  // 第二类健康打卡
+        "enable": false,  // true 为打开，false 为关闭
+        // 以下为数据修改，如果数据获取出现问题，请抓包修改或添加
+        "post_json":{
+          "address": "",
+          "stuNo": "",
+          //  百度坐标拾取：http://api.map.baidu.com/lbsapi/getpoint/index.html，可用来获取下面两个值
+          "longitude": "",  // 请对应 address 填写该值，否则老师那边会显示手动签到
+          "latitude": ""  // 请对应 address 填写该值，否则老师那边会显示手动签到
+        }
+      }
+    },
+    "campus_checkin": {  // 校内打卡
+      "enable": false,  // true 为打开，false 为关闭
+      "post_json": {  // 以下为数据修改，如果数据获取出现问题，请抓包修改对应项
+          "areaStr": "",
+          "deptStr": "",
+          "deptid": "",
+          "stuNo": "",
+          "username": "",
+          "userid": "",
+          "updatainfo": [
+            {
+              "propertyname": "temperature",
+              "value": "36.4"
+            },
+            {
+              "propertyname": "symptom",
+              "value": "无症状"
+            },
+          ]
+        }
+    },
+    "push": {  // 单人推送设置，若全部关闭，则使用 push.json 文件的配置，进行统一推送
+      "email": {  //  自定义邮箱推送，使用 qq 邮箱推送，就用 qq 邮箱的 smtp 服务地址和端口
+        "enable": true,  // true 为打开，false 为关闭
+        "smtp_address": "smtp.qq.com",  // stmp服务地址
+        "smtp_port": 465,  // stmp服务端口
+        "send_email": "***@qq.com",  // 发送邮箱的邮箱地址
+        "send_pwd": "****",  // 发送邮箱的邮箱授权码
+        "receive_email": "****@qq.com"  // 接收信息的邮箱地址
+      },
+      "wechat": {  //  Server酱·Turbo版：https://sct.ftqq.com/
+        "enable": false,  // true 为打开，false 为关闭
+        "send_key": ""  // Server 酱推送密钥
+      },
+      "qmsg": {  //  Qmsg酱：https://qmsg.zendee.cn/index.html
+        "enable": false,  // true 为打开，false 为关闭
+        "key": "",  // 推送所必须的 key
+        "type": "send",  //  send 私聊消息推送，group 群消息推送
+        "qq_num": ""  // 上面设置send，则填写接受推送的qq号，group则填写群号，根据官网要求群号需要邮箱申请才能推送
+      }
+    }
+  },
+]
