+++ requirements.txt
@@ -1,2 +1,5 @@
 pycryptodome==3.9.9
-requests==2.25.1
+requests~=2.25.1
+
+urllib3~=1.26.3
+json5~=0.9.5
