+++ setting.py
@@ -1,0 +1,9 @@
+import logging
+from utils.log import init_log
+
+__all__ = [
+    'log',
+
+]
+
+log = init_log(logging.INFO)
